feat: introduced JsonApiActionFilter
